Fixing crashing issue when multiple frames are in one audio packet.
When more than a single frame is part of a single compressed audio packet.
We push the packet back onto our queue, but remove the reference to it since
we processed part of it. The next time we try to use the partial packet, it
has been freed. Keep the reference on the packet when pushing it back to the
queue.

Also fix the timestamp for these packets. The frame might not have a pts
value, so keep the one from the packet.

Addresses issues #99 and #103