Set codec parameters from FLAC stream info in FLACSampleProvider (#274)
When FFmpeg is built without the FLAC decoder, some fields (e.g. channels, sample rate, bit rate) may not be set on the stream's codec parameters which can cause playback to fail.

This change updates the FLACSampleProvider to parse the FLAC stream info in the codec private data and set the stream's codec parameters accordingly.